Added @OneToMany Relations and Texto Controller, Model and Repository

+++ src/main/java/com/cespedesz07/nivel/NivelRepository.java
@@ -2,6 +2,6 @@ package com.cespedesz07.nivel;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface NivelRepository extends CrudRepository<Nivel, Long> {
+public interface NivelRepository extends CrudRepository<Nivel, Integer> {
 
 }
